+++ src/components/modules/ThemeSwitch.tsx
@@ -1,0 +1,28 @@
+import { useState, useEffect } from "react";
+import { useTheme } from "next-themes";
+import MoonIcon from "../icons/MoonIcon";
+import SunIcon from "../icons/SunIcon";
+
+const ThemeSwitch = () => {
+  const [mounted, setMounted] = useState(false);
+  const { theme, setTheme } = useTheme();
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  if (!mounted) {
+    return null;
+  }
+
+  return (
+    <button
+      className="cursor-pointer flex justify-center items-center"
+      onClick={() => setTheme(theme === "dark" ? "light" : "dark")}
+    >
+      {theme === "dark" ? <MoonIcon /> : <SunIcon />}
+    </button>
+  );
+};
+
+export default ThemeSwitch;