+++ src/types/global.d.ts
@@ -2,9 +2,23 @@ enum ConvertMode {
   JSONtoString,
   StringtoJSON,
 }
+
+enum Formats {
+  json,
+  string,
+}
+
 enum ColorScheme {
   Dark = "dark",
   Light = "light",
 }
 
-export { ConvertMode, ColorScheme };
+enum IconVariant {
+  beautify,
+  copy,
+  delete,
+  error,
+  uglify,
+}
+
+export { ConvertMode, ColorScheme, IconVariant, Formats };