+++ CHANGELOG.md
@@ -15,14 +15,38 @@ B: Nuevas características, nuevas funcionalidades y conjunto de grandes correcc
 C: Correcciones de bugs de la serie B.
 -->
 
+## [2.0.0] 02/08/2023
+
+### Added
+
+- Added error details.
+- Added button to toggle between JSON to String and String to JSON.
+- Added toolbar with the following buttons:
+  - Beautify: In JSON to String mode, formats JSON.
+  - Uglify: In JSON to String mode, removes indentation and extra spaces.
+  - Clear: Deletes content.
+  - Copy: Copies content to clipboard.
+- Added JSON linting.
+- Added About page.
+
+### Changed
+
+- Implemented new UI design.
+- Changed title from "Simple JSON Stringify" to "Simple JSON Stringifier".
+- Improved mobile version.
+
+### Fixed
+
+- Fixed bug that causes a flash of light mode when page loads and dark mode is enabled.
+
 ## [1.1.0] 08/04/2023
 
 ### Added
 
-- Add button to toggle light/dark mode.
+- Added button to toggle light/dark mode.
 
 ## [1.0.0] 08/04/2023
 
 ### Added
 
-- Add feature to convert a JSON formated string to a valid stringified JSON.
+- Added feature to convert a JSON formated string to a valid stringified JSON.